Add optional support for relative python paths
This is useful when copying a reference to use as an import in the
Python shell, where the Python path needs to be considered e.g.

   # src/path/to/file.py
   def some_function(): pass

   # python shell
   from path.to.file import some_function

This adds an optional global list of "paths" that will be removed
from the reference before other substitutions.

I thought it would be useful to add a test since this increases the
complexity of the script a bit.